Refactored  the package
com.api.pojo to com.api.request.model
