--- src/test/java/com/api/pojo/UserCredentials.java
@@ -1,6 +1,0 @@
-package com.api.pojo;
-
-public record UserCredentials(String username, String password)
-{
-
-}
