+++ src/test/java/com/api/request/model/CreateJobPayLoad.java
@@ -1,4 +1,4 @@
-package com.api.pojo;
+package com.api.request.model;
 
 import java.util.List;
