+++ src/test/java/com/api/test/CreateJobAPITest.java
@@ -8,11 +8,11 @@ import java.util.List;
 import org.testng.annotations.Test;
 
 import com.api.constants.Role;
-import com.api.pojo.CreateJobPayLoad;
-import com.api.pojo.Customer;
-import com.api.pojo.CustomerAddress;
-import com.api.pojo.CustomerProduct;
-import com.api.pojo.Problems;
+import com.api.request.model.CreateJobPayLoad;
+import com.api.request.model.Customer;
+import com.api.request.model.CustomerAddress;
+import com.api.request.model.CustomerProduct;
+import com.api.request.model.Problems;
 import com.api.utils.AuthTokenProvider;
 import com.api.utils.ConfigManager;
 import com.api.utils.SpecUtil;
