+++ src/test/java/com/api/test/LogInAPITest.java
@@ -5,7 +5,8 @@ import static org.hamcrest.Matchers.*;
 
 import org.testng.annotations.Test;
 
-import com.api.pojo.UserCredentials;
+import com.api.request.model.UserCredentials;
+
 import static com.api.utils.SpecUtil.*;
 
 import io.restassured.module.jsv.JsonSchemaValidator;
