+++ src/test/java/com/api/utils/AuthTokenProvider.java
@@ -6,7 +6,8 @@ import static org.hamcrest.Matchers.*;
 import static com.api.constants.Role.*;
 
 import com.api.constants.Role;
-import com.api.pojo.UserCredentials;
+import com.api.request.model.UserCredentials;
+
 import io.restassured.http.ContentType;
 
 public class AuthTokenProvider {
